o2t.ps1: fixed some edge cases re Path ordering
Consider cases where script is executed multiple times from within the
same shell, replacing a different Oracle JRE or JDK every time: This
requires refreshing the environment in between execution because the
Path changes with each iteration.
On Windows, it is possible to refresh the environment by leveraging the
native 'SendMessageTimeout' function. Not sure if this properly
documented anywhere, but it is mentioned in the ancient KB104011 article
(https://mskb.pkisolutions.com/kb/104011) and although Win Server 2003
was brand new when the article was published, the trick still works to
this day.