o2t.ps1: Inherit ExecutionPolicy from parent prompt.
In case that the script is started in a non-elevated prompt on a
system with a restrictive execution policy in place and the
execution policy for the system is relaxed for scope 'Process' only, the
admin window will fall back to the system-wide execution policy. As a
consequence, script execution in the admin shell might get blocked.

An easy fix is to get the current execution policy and pass it to the
admin shell via the '-ExecutionPolicy' command line switch.